training: move job specifications out of TrainingJob and into a separate file
Code cleanup, no functional changes
